les7: global styling & typography w/ styled components
The different thing in declaring CSS in JS is that it scoping things. Apply a specific style to a component it is scoped only for this use, unlike the Cascading Style Sheet meaning.
At navigation styles JS file there is NavStyles component, where at &:before{ } there are the styles for the skewed menu on the top, after the logo.