les8: visualizing route changes
NProgress is a JS library that visualize a page load actioln with a horizontal bar on the page.
Next JS library functions as a regulation on loading the page, on demand, means that the page elements is loading the time needed, and then have them local. This is in development mode. In production mode preFetch tags makes elements to load on page opening.