les15: setting up apollo client w/ react from mod: client side graphql
Redux is a state management system where you define actions and how those actions is handled and how the data goes in and out, and Apollo Client will replace the need for using something like Redux. Apollo does all the data management stuff that Redux store will do, as well as a whole bunch of other stuff. It will do the fetching of the data from the server side, the pushing of the data to your mutations. Also it will do all the caching for page requesting. The first time page requesting it will download it, so that to use it all the other times of request the same page. Apollo will also help on managing local states, even error catching in UI. Relay is a Facebook made library and is doing the same job with Apollo.
All you need to do in order to expose an Apollo client to a React app is just wrap it in an Apollo Provider. This means that you have to wrap everything after container component in rendering.

_app.js

<ApolloProvider client={this.props.apollo}>     </ApolloProvider>

export default withData(MyApp)