les17: creating items with mutations
The CreateItem component is going to have a local state inside. State is really helpful for the case that we just want to contain data locally within a component, and before send all the data to GraphQL API need to stored it somewhere. React idea is that a single source of truth, means the data to live in one spot.
By building an arrow function handleChange, ES6 is binding by default and you can use this to access instance properties. Handle change function allows to mirror multiple inputs to your state.