les34: permissions management
At the files of permissions.js of pages and the Permissiolns.js of components do these changes. We queried all the possible users, we are checking for displaying any errors, check all the possible permissions and loop over them. We take all the users and loop over them, and for each user we loop over the possible permissions and output the checkbox depending on if the user have this permission.