les35: updating permissions in local state
Using PropTypes in Permissions component and at the User class adding:
static propTypes = { }

At handlePermissionChange( ) take a copy of current permissions:
const updatedPermissions = [...this.state.permissions]

And then figure out if we need to remove or add this permission. If checkbox is checked add it in else remove it. All of these changes affects the React state of the UserPermissions component.