les38-39: create cart in react, local state queries & mutations
These are from the Module 6 learning about Shopping Cart

Create a new component Cart that holds the shopping goods in, and it is a sliding bar from the right on screen. Export TOGGLE_CART_MUTATION and import wherever it is need to trigger that mutation. So that, open the Nav.js and import Mutation from react-apollo and import TOGGLE_CART_MUTATION from Cart component. Last, import Cart.js at Header component.

At the withData.js file is where we created Apollo client, store and data saved local. For the Local data we have to read the cartOpen value from the cache, and write the cart State to the opposite.