les40: server side add to cart, shopping cart  module
Creating an AddToCart component and import that at Item.js in order to use it at the rendering. ADD_TO_CART_MUTATION at AddToCart.js  mutation addToCart($id: ID!) { }
The functionality of Adding to cart have the result of adding to database a cart item at the User that add it.