les41: displaying cart items and totals
In order to query the users cart we use the Nav component and we simply just need a user component and that will expose the currently logged in user. Inside that user we have info about the name the email and now we're gonna be able to query the users cart. So, at the Cart.js we use the User component and inside of it a function that takes destructure data and then destructure it one level further into a variable called me and pass it to the body of function that checks if me exists  if (!me) return null  With this component we wrap the Mutation component. So, we can further use me as the logged in user. So, at Supreme component we pass   <Supreme>{me.name}'s Cart</Supreme>  Also {me.cart.length} gives the number of items in the cart.
At User.js at the var CURRENT_USER_QUERY adding to the query the cart {id quantity}
Creating a new component CartItem and import it in Item component.
We use another utility function calcTotalPrice at the Cart.js that takes a user's cart and takes every single item and sort of mash each other. We starrt with 0 cents and then for each one we get a tally and the cart item and if there is a cart item then wea are going to return the tally times the new amount. We use it with formatMoney function that transform cents in money representation.