les43: optimistic response & cache updates w/ apollo
We are going at the RemoveFromCart.js and make some improvements to make the component work better. At the <Mutation mutation={REMOVE_FROM_CART_MUTATION}   update={this.update}> pass update function that have been defined before rendering. Update function gets called as soon as we get a response back from the server after a mutation has been performed.
	update = (cache, payload) =>
where passing apollo cache and payload which is the actual dump of info that returned from the server, in this case is the id. Then, inside update we do these:
 1. first read the cache
 2. remove that item from the cart
 3. write it back to the cache

 Because there is a split about a second to the response to the press of the button, we are going to make it a little faster, by applying the optimistic response method.