les44: animating our cart count component
We create a new component CartCount where we create a stylised div element Dot. In this we have CSS rules:	font-feature-settings: 'tnum';  font-variant-numeric: tabular-nums;
these rules make the numbers in this div to look the same width to each other.
We are importing this to the Nav.js and display it at the bar after MyCart button. The number that output coming from:
  count={me.cart.reduce((tally, cartItem) => tally + cartItem.quantity, 0)}
In order to show bold reaction of the MyCart number animating changes. The easiest way to transition from one component to the other component is by using the library react-transition-group. The result is an reaction to the Add to cart by bunching the number, flipping and rotate with the new number.