les46: cleaning up the render prop mess
Taking the already implemented Cart.js we see that we have User that queries the current user, that returns a mutation that exposes the mutation to toggle the cart open and close, that's gonna grab local state. To make this nested code prettier and easier to read we are using react-adopt.