les48: autocomplete w/ downshift from advanced ui module
The use of Downshift is going to be at Search.js with a render prop inside
  <Downshift>{( )=> ( )}</Downshift>
that gets after de-structure the getInputProps, getItemProps, isOpen, inputValue, hightlightedIndex.
At the search dropdown there is a Downshift functionality of highlight an item with arrow keys up and down, and with enter key select this item. In order to display the item title on search bar:
  <Downshift itemToString={item=> (item === null ? '' : item.title)}>
and to route to that item:
  <Downshift onChange={routeToItem( )}>
Use of the   resetIdCounter()  first thing of rendering solves the Downshift  Warning: Prop `aria-labeled` did not match. An import of this method needed to use it.