les54-55: unit testing 101 w/ jest and enzyme introduction
Jest is used by Facebook to test all JavaScript code including React applications. One of Jest's philosophies is to provide an integrated "zero-configuration" experience.
Enzyme is a JavaScript Testing utility for React that makes it easier to assert, manipulate, and traverse your React Components' output.
Enzyme's API is meant to be intuitive and flexible by mimicking jQuery's API for DOM manipulation and traversal.
Tests will be in a __tests__ folder, or else name test files with a .spec.js or .test.js extension. Jest   describe('  ', ()=> { it('  ', ()=> { expect( ).toEqual( ); }); })   is a global method that first takes a description of what it is doing and at the arrow function returns the   test()  or the it()  Also as it() others are:
skip it   xit('  ', ()=> { expect( ).toEqual( ); })   and that will skip it
focus it  fit('  ', ()=> { expect( ).toEqual( ); })   and will focus only on it
As about the formatMoney.js in order to test it, we create a new test file formatMoney.test.js where   import formatMoney from '../lib/formatMoney';   and in this test suite   describe(' ', ()=> {  })   we put some   expect(formatMoney( )).toEqual('  ')   so to check that the function works fine.
At package.json and at the script  "test": "NODE_ENV=test jest --watch"