les56: mocking 101
Mocking is when a part of the application connected to another part outside of it, like an external API, in order to test the calls to this API you have to mock API calls, so that to pretend that function of fetching data and returning something. So your unit test is very self-contained. To do that we create a new test file mocking.test.js and   describe('  ', ()=> {it('  ', ()=> { const fetchDogs= jest.fn() })})   so we have a mocking function fetchDogs()     http://jestjs.io/docs/en/mock-funcion-api