les57: first tests and shallow rendering
We are going to take --- component that it past in data via props and display them. So, the test of it have to do with render it and check that all these things comes with data are working as planned. Test is checking that the description exists, the title is there, the image exists and the price is being formatted properly.