les58: snapshot testing
It is called snapshot because you can take a photo of a component and if that were to ever change, then it's going to alert that test are failing. So that, a snapshot is an exact representation of the component at the time of the snapshot taken.

Mounting makes it actually run like a browser environment, whereas shallow only goes one level deep.