les59: testing and mocking apollo queries
Taking the SingleItem component, it requires a query that's past the id of an item and as it comes back it will render an error or it will render the loading or no-item-found. In order to test all of the elements of the component we will use  import {mount} from 'enzyme'  Also toJSON, wait, MockedProvider, fakeItem.
The mocking is in the describe('  ', ()=> {it ('  ', async ()=> { const mocks= [ ] }) })
 - When someone makes a request with this query and variable combo
 - return this fake data (mocked data)
We will not take a whole snapshot because it is a lot of lines of elements. So we take a snapshot of the elements we want to test.